Back down C++20 experiment. Ahh well. (#7656)
This reverts commit 7e19415754d360c74842a8d967d652e9d981bd7d.

+++ src/surge-xt/gui/widgets/XMLConfiguredMenus.h
@@ -109,7 +109,7 @@ struct XMLMenuPopulator
         {
             idx += dir;
             if (idx < 0)
-                idx = (int)allPresets.size() - 1;
+                idx = allPresets.size() - 1;
 
             if (idx >= (int)allPresets.size())
                 idx = 0;
